Add files via upload
Given an N×N chessboard, with two pieces, a black pawn and a white pawn randomly placed on the chessboard. What is the minimum number of steps required by the white pawn to reach the black pawn under the following assumptions.

a. The pawns are always inside the board.
b. The pawns can move in either if the 8 directions.
c. The pawns can move only 1 step at a time